requirements.txt
aiofiles==24.1.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.1
openai-whisper==20231117
priority==2.0.0
Quart==0.19.6
regex==2024.5.15
requests==2.32.3
sympy==1.13.0
taskgroup==0.0.0a4
tiktoken==0.7.0
tomli==2.0.1
torch==2.2.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2